# PERF: Add partial index on reviewables for topic view